workflow: separate the e2e tests that requires real dependency installation
